Address comments, clear up test cases
