+++ tools/testdata/qwen2.5.gotmpl
