+++ tools/tools_utils.go
@@ -23,7 +23,7 @@ import (
 //   - bool: Whether a ".ToolCalls" condition was found in the template
 func extractToolCallsFormat(tmpl *gotmpl.Template) (string, bool) {
 	if tmpl == nil || tmpl.Tree == nil {
-		slog.Debug("TextAfterToolCalls: template or tree is nil")
+		slog.Debug("template or tree is nil")
 		return "", false
 	}
 
@@ -74,10 +74,6 @@ func extractToolCallsFormat(tmpl *gotmpl.Template) (string, bool) {
 			default:
 				// Continue to next node
 				continue
-			}
-
-			if found {
-				return
 			}
 		}
 	}
@@ -225,82 +221,4 @@ func collect(obj any) []map[string]any {
 	}
 
 	return all
-}
-
-// parseJSONToolCalls attempts to parse a JSON string into a slice of ToolCalls.
-//
-// Parameters:
-//   - s: The string to parse
-//   - name: The field name from template that identifies the tool call name
-//   - arguments: The field name from template that identifies the tool call arguments
-//
-// Returns:
-//   - []api.ToolCall: The parsed tool calls if successful
-//   - error: ErrAccumulateMore if braces unbalanced, ErrInvalidToolCall if invalid, or nil if successful
-func parseJSONToolCalls(s string, name, arguments string) ([]api.ToolCall, error) {
-	// Check for balanced braces before attempting to parse
-	braceCount := 0
-	startIndex := -1
-	var rawToolCalls []string
-
-	for i, c := range s {
-		if c == '{' {
-			braceCount++
-			if startIndex == -1 {
-				startIndex = i
-			}
-		} else if c == '}' {
-			braceCount--
-			if braceCount == 0 {
-				rawToolCalls = append(rawToolCalls, s[startIndex:i+1])
-				startIndex = -1
-			}
-		}
-		// Negative means we have an extra closing brace
-		if braceCount < 0 {
-			return nil, errInvalidToolCall
-		}
-	}
-
-	// If braces aren't balanced, need more input
-	if braceCount > 0 {
-		slog.Debug("unbalanced braces detected", "input", s)
-		return nil, errAccumulateMore
-	}
-
-	// Attempt full unmarshal of the JSON
-	var toolCalls []api.ToolCall
-	for _, rawToolCall := range rawToolCalls {
-		var resp any
-		if err := json.Unmarshal([]byte(rawToolCall), &resp); err != nil {
-			continue
-		}
-
-		// Collect nested objects that could contain tool calls
-		objs := collect(resp)
-		if len(objs) == 0 {
-			continue
-		}
-
-		// Extract tool calls from objects
-		for _, kv := range objs {
-			n, nok := kv[name].(string)
-			a, aok := kv[arguments].(map[string]any)
-			if nok && aok {
-				toolCalls = append(toolCalls, api.ToolCall{
-					Function: api.ToolCallFunction{
-						Name:      n,
-						Arguments: a,
-					},
-				})
-			}
-		}
-	}
-
-	// Valid JSON, no tool calls found
-	if len(toolCalls) == 0 {
-		return nil, errInvalidToolCall
-	}
-
-	return toolCalls, nil
 }
